requirements.txt
fastapi==0.111.1
psycopg2-binary
pyluach==2.2.0
pytest==7.3.1
SQLAlchemy==2.0.31
uvicorn==0.30.3
python-dotenv
requests